Begin port of Raku version
This is a port to Perl of an (unpublished) Raku program for downloading
purchases from Nugs.net. The downloads used to run on platform called
"muletracks", hence the name. The old version does not work anymore due
to the platform change, but I want to keep the name.

The gears are all there already, but a convenient script has be written
around them still. Documentation and tests are missing.